Fixed Copy scroll issues: Destruction of entire stack bug + disallow copying spells not found in wizard repertoire

+++ TemplePlus/python/python_dispatcher.cpp
@@ -402,7 +402,7 @@ PYBIND11_PLUGIN(tp_dispatcher){
 			.def_readwrite("caster_level", &SpellPacketBody::casterLevel)
 			.def("get_target",[](SpellPacketBody &pkt, int idx)->objHndl
 			{
-				if (idx < pkt.targetCount)
+				if (idx < (int)pkt.targetCount)
 					return pkt.targetListHandles[idx];
 				return objHndl::null;
 			} )
